Feat food

See merge request s09-final/S09P31B205!30

Former-commit-id: 1cc12754d7c5ff56ccc2fcde6031e3c00d8b176d

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/hotel/dto/request/AttractionCoordinateRequestDto.java
@@ -9,8 +9,8 @@ import java.math.BigDecimal;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 public class AttractionCoordinateRequestDto {
-    private BigDecimal latitude;
-    private BigDecimal longitude;
-    private Long distance;
+    private Double latitude;
+    private Double longitude;
+    private Integer distance;
     private String sorted;
 }
